Rename FormBrowseController -> GpgInfoProvider
