+++ src/app/GitUI/CommandsDialogs/FormBrowse.cs
@@ -7,7 +7,7 @@ using ConEmu.WinForms;
 using GitCommands;
 using GitCommands.Config;
 using GitCommands.Git;
-using GitCommands.Gpg;
+using GitCommands.Git.Gpg;
 using GitCommands.Submodules;
 using GitCommands.Utils;
 using GitExtensions.Extensibility;
@@ -208,7 +208,7 @@ namespace GitUI.CommandsDialogs
         private readonly SplitterManager _splitterManager;
         private readonly GitStatusMonitor _gitStatusMonitor;
         private readonly FormBrowseMenus _formBrowseMenus;
-        private readonly IFormBrowseController _controller;
+        private readonly IGpgInfoProvider _controller;
         private readonly ICommitDataManager _commitDataManager;
         private readonly IAppTitleGenerator _appTitleGenerator;
         private readonly IAheadBehindDataProvider? _aheadBehindDataProvider;
@@ -306,7 +306,7 @@ namespace GitUI.CommandsDialogs
             UICommands.PostRepositoryChanged += UICommands_PostRepositoryChanged;
             UICommands.BrowseRepo = this;
 
-            _controller = new FormBrowseController(new GitGpgController(() => Module));
+            _controller = new GpgInfoProvider(new GitGpgController(() => Module));
             _commitDataManager = new CommitDataManager(() => Module);
 
             _submoduleStatusProvider = commands.GetRequiredService<ISubmoduleStatusProvider>();
